Add support for selective testing for Xcode projects with local packages
