+++ Sources/TuistLoader/Loaders/ManifestLoader.swift
@@ -6,6 +6,7 @@ import ProjectDescription
 import ServiceContextModule
 import TuistCore
 import TuistSupport
+import XcodeGraph
 
 public enum ManifestLoaderError: FatalError, Equatable {
     case projectDescriptionNotFound(AbsolutePath)
