+++ Sources/TuistLoader/SwiftPackageManager/PackageInfo+graphPlatform.swift
@@ -1,6 +1,7 @@
 import Foundation
 import ProjectDescription
 import TuistSupport
+import XcodeGraph
 
 extension PackageInfo.Platform {
     func destinations() throws -> ProjectDescription.Destinations {
