+++ Tests/TuistKitTests/Services/Plugin/PluginArchiveServiceTests.swift
@@ -3,6 +3,7 @@ import ServiceContextModule
 import TuistLoader
 import TuistSupport
 import TuistSupportTesting
+import XcodeGraph
 import XCTest
 @testable import TuistKit
