+++ Tests/TuistLoaderTests/Loaders/PackageInfoLoaderTests.swift
@@ -2,6 +2,7 @@ import TSCUtility
 import TuistCore
 import TuistSupport
 import TuistSupportTesting
+import XcodeGraph
 import XCTest
 @testable import TuistLoader
