+++ fixtures/xcode_project_with_packages_and_tests/.gitignore
@@ -1,0 +1,4 @@
+!App.xcodeproj
+!App.xcodeproj/project.xcworkspace
+App.xcodeproj/project.xcworkspace/xcuserdata
+App.xcodeproj/project.xcworkspace/contents.xcworkspacedata
