+++ fixtures/xcode_project_with_packages_and_tests/App/LibraryA/Sources/LibraryA/LibraryA.swift
@@ -1,0 +1,3 @@
+import LibraryB
+
+struct LibraryA {}
