+++ fixtures/xcode_project_with_packages_and_tests/App/LibraryA/Tests/LibraryATests/LibraryATests.swift
@@ -1,0 +1,6 @@
+import Testing
+@testable import LibraryA
+
+@Test func libraryAExample() async throws {
+    // Write your test here and use APIs like `#expect(...)` to check expected conditions.
+}
