+++ fixtures/xcode_project_with_packages_and_tests/App/LibraryB/Package.swift
@@ -1,0 +1,26 @@
+// swift-tools-version: 6.0
+// The swift-tools-version declares the minimum version of Swift required to build this package.
+
+import PackageDescription
+
+let package = Package(
+    name: "LibraryB",
+    products: [
+        // Products define the executables and libraries a package produces, making them visible to other packages.
+        .library(
+            name: "LibraryB",
+            targets: ["LibraryB"]
+        ),
+    ],
+    targets: [
+        // Targets are the basic building blocks of a package, defining a module or a test suite.
+        // Targets can depend on other targets in this package and products from dependencies.
+        .target(
+            name: "LibraryB"
+        ),
+        .testTarget(
+            name: "LibraryBTests",
+            dependencies: ["LibraryB"]
+        ),
+    ]
+)
