+++ fixtures/xcode_project_with_packages_and_tests/App/LibraryB/Sources/LibraryB/LibraryB.swift
@@ -1,0 +1,2 @@
+// The Swift Programming Language
+// https://docs.swift.org/swift-book
