+++ fixtures/xcode_project_with_packages_and_tests/App/LibraryB/Tests/LibraryBTests/LibraryBTests.swift
@@ -1,0 +1,6 @@
+import Testing
+@testable import LibraryB
+
+@Test func example() async throws {
+    // Write your test here and use APIs like `#expect(...)` to check expected conditions.
+}
