+++ fixtures/xcode_project_with_packages_and_tests/AppTests/AppTests.swift
@@ -1,0 +1,8 @@
+import Testing
+@testable import App
+
+struct AppTests {
+    @Test func example() async throws {
+        // Write your test here and use APIs like `#expect(...)` to check expected conditions.
+    }
+}
