+++ fixtures/xcode_project_with_packages_and_tests/Tuist.swift
@@ -1,0 +1,9 @@
+import ProjectDescription
+
+let config = Config(
+    fullHandle: "tuist/xcode_project_with_tests",
+    url: "https://canary.tuist.dev",
+    generationOptions: .options(
+        optionalAuthentication: true
+    )
+)
