+++ mise.toml
@@ -1,5 +1,5 @@
 [tools]
-tuist = "4.41.0"
+tuist = "4.42.0"
 swiftlint = "0.54.0"
 swiftformat = "0.53.3"
 pnpm = "8.15.6"
